astro-engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyswisseph>=2.10.3.2
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dateutil>=2.9.0
httpx>=0.27.0
reportlab>=4.2.0
python-multipart>=0.0.9
